cdk/requirements.txt
aws-cdk.core==1.107.0
aws-cdk.aws-iam==1.107.0
aws-cdk.aws-lambda==1.107.0
aws-cdk.aws-sns==1.107.0
aws-cdk.aws_sns_subscriptions==1.107.0
